empty_example implemented and looks good! #68

+++ power_perceiver/load_raw/data_sources/raw_gsp_data_source.py
@@ -76,13 +76,13 @@ class RawGSPDataSource(
         gsp_id_to_shape = gsp_id_to_shape.loc[gsp_pv_power_mw_ds.gsp_id]
 
         data_array = _put_gsp_data_into_an_xr_dataarray(
-            gsp_pv_power_mw=gsp_pv_power_mw_ds.generation_mw.data,
+            gsp_pv_power_mw=gsp_pv_power_mw_ds.generation_mw.data.astype(np.float32),
             time_utc=gsp_pv_power_mw_ds.datetime_gmt.data,
             gsp_id=gsp_pv_power_mw_ds.gsp_id.data,
             # TODO: Try using `gsp_id_to_shape.geometry.envelope.centroid`. See issue #76.
-            x_osgb=gsp_id_to_shape.geometry.centroid.x,
-            y_osgb=gsp_id_to_shape.geometry.centroid.y,
-            capacity_mwp=gsp_pv_power_mw_ds.installedcapacity_mwp.data,
+            x_osgb=gsp_id_to_shape.geometry.centroid.x.astype(np.float32),
+            y_osgb=gsp_id_to_shape.geometry.centroid.y.astype(np.float32),
+            capacity_mwp=gsp_pv_power_mw_ds.installedcapacity_mwp.data.astype(np.float32),
         )
 
         del gsp_id_to_shape, gsp_pv_power_mw_ds
